[TS-337] Refactor: create habit directory refactoring
