+++ src/components/Habit/HabitTip/HabitTip.style.ts
